many errors in queries
I had to fix them so that I could add sample records in discount table
to display deals in home page.

+++ sql/create.sql
@@ -82,7 +82,7 @@ create table discount(
 	cur_date date not null,
 	start_date date not null,
 	end_date date not null,
-	rate int not null,
+	rate decimal not null,
 	constraint chk_rate check (rate > 0 and rate < 100),
 	primary key (id),
 	foreign key (room_type_id) references room_type(id),
